改

+++ src/views/project/projectList.vue
@@ -1,10 +1,16 @@
 <template>
     <div class="warnWrapper">
         <el-table :data="tableData" style="width: 100%">
-            <el-table-column  prop="projectName" label="项目名称" width="180"></el-table-column>
-            <el-table-column  prop="projectIntroduce" label="项目介绍" width="180"></el-table-column>
-            <el-table-column  prop="projectKeyword" label="项目关键字" width="180"></el-table-column>
-            <el-table-column  prop="state" label="审核状态" width="180"></el-table-column>
+            <el-table-column  prop="projectName" label="项目名称" width="200"></el-table-column>
+            <el-table-column  prop="projectIntroduce" label="项目介绍"></el-table-column>
+            <el-table-column  prop="projectKeyword" label="项目关键字"  width="180"></el-table-column>
+            <el-table-column label="审核状态" width="180" class-name="checkState">
+                <template slot-scope="scope">
+                    <span class="over" v-if="scope.row.state == 'N'">通过</span>
+                    <span class="fail" v-if="scope.row.state == 'F'">未通过</span>
+                    <span class="wait" v-if="scope.row.state == 'W'">待审核</span>
+                </template>
+            </el-table-column>
             <el-table-column  prop="createDate" label="日期" width="180"></el-table-column>
             <el-table-column  label="操作" width="180">
                 <template slot-scope="scope">
@@ -288,6 +294,7 @@ export default {
             }
             listCompanyProject(myData)
             .then(res=>{
+                this.total = res.result.total
                 this.tableData = res.result.list
             })
         },
@@ -300,15 +307,21 @@ export default {
             })
         },
         handleDel(id){
-            let comName = JSON.parse(sessionStorage.getItem("user")).comName
-            let myData = {
-                companyProjectId:id,
-                comName:comName
-            }
-            deleteProject(myData)
-            .then(res => {
-                this.getData()
-            })
+            this.$confirm('此操作将删除该项目, 是否继续?', '提示', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+                }).then(() => {
+                    let comName = JSON.parse(sessionStorage.getItem("user")).comName
+                    let myData = {
+                        companyProjectId:id,
+                        comName:comName
+                    }
+                    deleteProject(myData)
+                    .then(res => {
+                        this.getData()
+                    })
+                })
         },
         handleSizeChange(val){
             this.pageSize = val
